+++ parser/src/processors/type_processors/class_call.rs
@@ -1,3 +1,10 @@
+/*
+
+
+
+DEPRECATED
+
+
 use crate::parser;
 use crate::processors::{type_processors, value_processor};
 use crate::syntax::{definers, types, variable};
@@ -8,17 +15,15 @@ use alloc::vec::Vec;
 use ellie_core::error;
 use ellie_core::{defs, utils};
 
-pub fn collect_class_call<F,E>(
-    parser: parser::Parser<F,E>,
+pub fn collect_class_call<F, E>(
+    parser: parser::Parser<F, E>,
     itered_data: &mut variable::VariableCollector,
     errors: &mut Vec<error::Error>,
     letter_char: &str,
     next_char: &str,
     last_char: &str,
 ) where
-        F: FnMut(ellie_core::com::Message) + Clone + Sized,
-    
-    
+    F: FnMut(ellie_core::com::Message) + Clone + Sized,
 {
     if let types::Types::ConstructedClass(ref mut class_call_data) = itered_data.data.value {
         if !class_call_data.keyword_collected {
@@ -32,7 +37,7 @@ pub fn collect_class_call<F,E>(
             } else if (letter_char == " " && class_call_data.keyword_index == 0)
                 || letter_char != " "
             {
-                                errors.push(error::Error {
+                errors.push(error::Error {
                     path: parser.options.path.clone(),
                     scope: "function_call_processor".to_owned(),
                     debug_message: "92dad599da8f896b7f972942e8337c06".to_owned(),
@@ -91,7 +96,7 @@ pub fn collect_class_call<F,E>(
                     class_call_data.name_collected = true;
                 }
             } else if letter_char != " " {
-                                errors.push(error::Error {
+                errors.push(error::Error {
                     path: parser.options.path.clone(),
                     scope: "function_call_processor".to_owned(),
                     debug_message: "0d27b8f6b7a72f5bb39e259bdfb00f7a".to_owned(),
@@ -176,8 +181,8 @@ pub fn collect_class_call<F,E>(
                     class_call_data.data.params[last_entry - 1].pos.range_end = parser.pos;
                 }
 
-                let fn_exists = parser.resolve_class_call(class_call_data.clone());
-                if let Some(type_errors) = fn_exists {
+                let class_exists = parser.resolve_class_call(class_call_data.clone());
+                if let Some(type_errors) = class_exists {
                     for error in type_errors {
                         errors.push(error);
                     }
@@ -342,14 +347,16 @@ pub fn collect_class_call<F,E>(
                             },
                         }
                     }
-                    types::Types::ConstructedClass(match_data) => types::class_call::ClassCallParameter {
-                        value: types::Types::ConstructedClass(match_data),
-                        pos: if last_entry == 0 {
-                            defs::Cursor::default()
-                        } else {
-                            class_call_data.data.params[last_entry - 1].pos
-                        },
-                    },
+                    types::Types::ConstructedClass(match_data) => {
+                        types::class_call::ClassCallParameter {
+                            value: types::Types::ConstructedClass(match_data),
+                            pos: if last_entry == 0 {
+                                defs::Cursor::default()
+                            } else {
+                                class_call_data.data.params[last_entry - 1].pos
+                            },
+                        }
+                    }
                     types::Types::Negative(match_data) => types::class_call::ClassCallParameter {
                         value: types::Types::Negative(match_data),
                         pos: if last_entry == 0 {
@@ -406,14 +413,13 @@ pub fn collect_class_call<F,E>(
                 }
             }
         } else if letter_char == "." {
-            itered_data.data.value =
-                types::reference_type::ReferenceTypeCollector {
-                    data: types::reference_type::ReferenceType {
-                        reference: Box::new(itered_data.data.value.clone()),
-                        chain: Vec::new(),
-                    },
-                    on_dot: false,
-                };
+            itered_data.data.value = types::reference_type::ReferenceTypeCollector {
+                data: types::reference_type::ReferenceType {
+                    reference: Box::new(itered_data.data.value.clone()),
+                    chain: Vec::new(),
+                },
+                on_dot: false,
+            };
             type_processors::reference::collect_reference(
                 parser,
                 itered_data,
@@ -479,3 +485,4 @@ pub fn collect_class_call<F,E>(
         }
     }
 }
+*/