+++ runtime/src/lib.rs
@@ -32,4 +32,3 @@ extern crate std;
 
 pub mod heap;
 pub mod runtime;
-pub mod thread;