+++ runtime/src/thread.rs
@@ -1,1 +1,5 @@
 
+pub struct Thread<F> {
+    pub name: String,
+    pub com: F
+}